plan improvements
