+++ src/HotChocolate/Fusion-vnext/src/Fusion.Execution/Execution/Nodes/ExecutionNodeResult.cs
@@ -1,5 +1,6 @@
 using System.Collections.Immutable;
 using System.Diagnostics;
+using HotChocolate.Language;
 
 namespace HotChocolate.Fusion.Execution.Nodes;
 
@@ -9,4 +10,5 @@ internal sealed record ExecutionNodeResult(
     ExecutionStatus Status,
     TimeSpan Duration,
     Exception? Exception,
-    ImmutableArray<ExecutionNode> DependentsToExecute);
+    ImmutableArray<ExecutionNode> DependentsToExecute,
+    ImmutableArray<VariableValues> VariableValueSets);
