+++ src/HotChocolate/Fusion-vnext/src/Fusion.Execution/Execution/Nodes/ExecutionNodeTrace.cs
@@ -1,3 +1,6 @@
+using System.Collections.Immutable;
+using HotChocolate.Language;
+
 namespace HotChocolate.Fusion.Execution.Nodes;
 
 public sealed class ExecutionNodeTrace
@@ -9,4 +12,6 @@ public sealed class ExecutionNodeTrace
     public required TimeSpan Duration { get; init; }
 
     public required ExecutionStatus Status { get; init; }
+
+    public ImmutableArray<VariableValues> VariableSets { get; init; }
 }
